requirements.txt
streamlit
numpy
pandas
Pillow
colormath
scikit-image
